+++ src/main/java/com/itilria/altograce/repository/UserRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.Query;
 import com.itilria.altograce.domain.User;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long>{
-    Optional<User> findByAuthid_Id(long id);
+public interface UserRepository extends JpaRepository<User, Integer>{
+    Optional<User> findByAuthid_Id(int id);
     @Query("SELECT u FROM User u WHERE u.id_passport = ?1")
     Optional<User> findByPassport(String id_passport);
 }
