+++ src/main/java/com/itilria/altograce/repository/clientrepository/ClientBillingRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.itilria.altograce.domain.client.ClientBilling;
 
-public interface ClientBillingRepository extends JpaRepository<ClientBilling, Long>{
-    List<ClientBilling> findByPrimaryClient_Id(long id);
+public interface ClientBillingRepository extends JpaRepository<ClientBilling, Integer>{
+    List<ClientBilling> findByPrimaryClient_Id(int id);
 }
