+++ src/main/java/com/itilria/altograce/repository/clientrepository/ClientSettingsRepository.java
@@ -7,6 +7,6 @@ import com.itilria.altograce.domain.client.ClientSettings;
 import java.util.Optional;
 import java.util.List;
 
-public interface ClientSettingsRepository extends JpaRepository<ClientSettings, Long>{
-    Optional<ClientSettings> findByCompany_Id(long comId);
+public interface ClientSettingsRepository extends JpaRepository<ClientSettings, Integer>{
+    Optional<ClientSettings> findByCompany_Id(int comId);
 }
